+++ lib/cucumber/configuration.rb
@@ -62,6 +62,14 @@ module Cucumber
       @options[:dry_run]
     end
 
+    def publish_enabled?
+      @options[:publish_enabled]
+    end
+
+    def publish_quiet?
+      @options[:publish_quiet]
+    end
+
     def fail_fast?
       @options[:fail_fast]
     end
@@ -256,6 +264,7 @@ module Cucumber
         strict: Cucumber::Core::Test::Result::StrictConfiguration.new,
         require: [],
         dry_run: false,
+        publish_quiet: false,
         fail_fast: false,
         formats: [],
         excludes: [],